refactor(codegen): split if-branch component emission into branch_component
Move `generate_if_branch_component` (the largest if-branch-shape emitter) out of
`v_if/branch.rs` into a new `branch_component` module. It is self-contained —
only imported helpers, no calls back into branch.rs — so the extraction is a
clean code movement.

Pure code movement — no behavior change. branch.rs shrinks 628 -> 390;
branch_component.rs is 270. Verified across vize_atelier_dom, vize_atelier_sfc
(528), and vize_atelier_ssr (76); clippy clean.
